modified:   src/app/test/echo/udp_echo_client.c 	modified:   src/app/test/main.c 	modified:   src/net/net/net_cfg.h 	modified:   src/net/net/net_err.h 	modified:   src/net/net/sock.h 	modified:   src/net/net/socket.h 	modified:   src/net/net/tools.h 	modified:   src/net/net/udp.h 	modified:   src/net/src/exmsg.c 	modified:   src/net/src/ipv4.c 	modified:   src/net/src/raw.c 	modified:   src/net/src/tools.c 	modified:   src/net/src/udp.c

+++ src/net/net/socket.h
@@ -66,7 +66,7 @@ struct x_sockaddr_in
 {
     uint8_t sin_len;
     uint8_t sin_family;
-    uint8_t sin_port;
+    uint16_t sin_port;
     struct x_in_addr sin_addr;
     char sin_zero[8];               // 填充字段，一般全为0
 };
